refactor!: Move implementations next to API
BREAKING CHANGE: Some bounds and `ErrorKind`s were changed
